refactor: reorganize imports and rename generatores filename to controller_generator
